MainMenu Update
matchStart, campaign, option, Exit
