+++ UnityProject/Assets/Script/Campaign/CampaignButtonControllScript.cs
@@ -1,0 +1,15 @@
+﻿using UnityEngine;
+using System.Collections;
+
+public class CampaignButtonControllScript : MonoBehaviour {
+
+	// Use this for initialization
+	void Start () {
+	
+	}
+	
+    public void CampaignBackButton()
+    {
+        Application.LoadLevel("MainMenu");
+    }
+}
