+++ UnityProject/Assets/Script/Match/MatchButtonControllerScript.cs
@@ -1,0 +1,13 @@
+﻿using UnityEngine;
+using System.Collections;
+
+public class MatchButtonControllerScript : MonoBehaviour {
+
+	
+	
+    public void MainButtonInvoke()
+    {
+        Application.LoadLevel("MainMenu");
+    }
+	
+}
