--- UnityProject/Assets/Script/MatchButtonControllerScript.cs
@@ -1,13 +1,0 @@
-﻿using UnityEngine;
-using System.Collections;
-
-public class MatchButtonControllerScript : MonoBehaviour {
-
-	
-	
-    public void MainButtonInvoke()
-    {
-        Application.LoadLevel("MainMenu");
-    }
-	
-}
