--- UnityProject/Assets/Script/QueueButtonRemain.cs
@@ -1,15 +1,0 @@
-﻿using UnityEngine;
-using System.Collections;
-
-public class QueueButtonRemain : MonoBehaviour {
-
-	// Use this for initialization
-	void Start () {
-        DontDestroyOnLoad(this.gameObject);
-	}
-	
-	// Update is called once per frame
-	void Update () {
-	
-	}
-}
