+++ UnityProject/ProjectSettings/GraphicsSettings.asset
@@ -22,7 +22,7 @@ GraphicsSettings:
   - {fileID: 10782, guid: 0000000000000000f000000000000000, type: 0}
   m_PreloadedShaders: []
   m_ShaderSettings:
-    useScreenSpaceShadows: 1
+    useScreenSpaceShadows: 0
   m_BuildTargetShaderSettings: []
   m_LightmapStripping: 0
   m_FogStripping: 0
